+++ tests-data/union-repeated-aliased-type/out.gen.ts
@@ -1,0 +1,60 @@
+import * as bare from "@bare-ts/lib"
+
+export type f32 = number
+export type f64 = number
+export type i8 = number
+export type i16 = number
+export type i32 = number
+export type i64 = bigint
+export type i64Safe = number
+export type int = bigint
+export type intSafe = number
+export type u8 = number
+export type u16 = number
+export type u32 = number
+export type u64 = bigint
+export type u64Safe = number
+export type uint = bigint
+export type uintSafe = number
+
+export type Y = u8
+
+export const readY = bare.readU8
+
+export const writeY = bare.writeU8
+
+export type X = 
+    | { readonly tag: 0; readonly val: u8 }
+    | { readonly tag: 1; readonly val: Y }
+
+export function readX(bc: bare.ByteCursor): X {
+    const offset = bc.offset
+    const tag = bare.readU8(bc)
+    switch (tag) {
+        case 0: {
+            const val = (bare.readU8)(bc)
+            return { tag, val }
+        }
+        case 1: {
+            const val = (readY)(bc)
+            return { tag, val }
+        }
+        default: {
+            bc.offset = offset
+            throw new bare.BareError(offset, "invalid tag")
+        }
+    }
+}
+
+export function writeX(bc: bare.ByteCursor, x: X): void {
+    const tag = x.tag;
+    bare.writeU8(bc, tag)
+    switch (tag) {
+        case 0:
+            (bare.writeU8)(bc, x.val)
+            break
+        case 1:
+            (writeY)(bc, x.val)
+            break
+    }
+}